+++ src/analysis/parsetree_utils.mli
@@ -6,3 +6,6 @@ open Parsetree
 type nonrec constant_desc = constant_desc
 
 val constant_desc : constant -> constant_desc
+
+(** Filter parsetree attributes which are prefixed by ["merlin."] in given expression. *)
+val expr_remove_merlin_attributes : expression -> expression