+++ CHANGES.md
@@ -1,6 +1,8 @@
 unreleased
 ==========
+
   + merlin library
+    - Implement new refactor-extract-region command for extracting region to a fresh let binding (#1948)
     - Fix `merlin_reader` for OpenBSD (#1956)
     - Improve recovery of mutually recursive definitions (#1962, #1963, fixes #1953)
   + vim plugin